internal/usergroup: move to daemon/internal
