+++ daemon/daemon_unix.go
@@ -30,10 +30,10 @@ import (
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/initlayer"
 	"github.com/docker/docker/daemon/internal/libcontainerd/remote"
+	"github.com/docker/docker/daemon/internal/usergroup"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/internal/nlwrap"
 	"github.com/docker/docker/internal/otelutil"
-	"github.com/docker/docker/internal/usergroup"
 	"github.com/docker/docker/libnetwork"
 	nwconfig "github.com/docker/docker/libnetwork/config"
 	"github.com/docker/docker/libnetwork/drivers/bridge"
